チャプター&レッスン画面にチャプター進捗などを表示（受講生側）：Resourseファイル作成
